+++ vms/saevm/sae/rpc/stateful.go
@@ -124,7 +124,7 @@ func (b *backend) stateAtBlock(ctx context.Context, num uint64) (*state.StateDB,
 		return nil, nil, err
 	}
 
-	if len(toReexec) > 0 {
+	if len(toReexec) > 0 && b.replaySlots != nil {
 		select {
 		case b.replaySlots <- struct{}{}:
 			defer func() { <-b.replaySlots }()